Add Ingredient Repository and implementation methods

+++ CookTheWeek.Data.Repositories/FavouriteRecipeRepository.cs
@@ -3,10 +3,10 @@
     using Microsoft.EntityFrameworkCore;
 
     using CookTheWeek.Common.Exceptions;
+    using CookTheWeek.Common.HelperMethods;
     using CookTheWeek.Data.Models;
 
     using static CookTheWeek.Common.ExceptionMessagesConstants.RecordNotFoundExceptionMessages;
-    using CookTheWeek.Common.HelperMethods;
 
     public class FavouriteRecipeRepository : IFavouriteRecipeRepository
     {
@@ -16,9 +16,8 @@
         {
             this.dbContext = dbContext;
         }
-        /// <summary>
-        ///  Returns true if the recipe is liked by the user
-        /// </summary>
+        
+        /// <inheritdoc/>
         public async Task<bool> GetByIdAsync(string userId, string recipeId)
         {
             return await this.dbContext.FavoriteRecipes
@@ -26,9 +25,7 @@
                                 GuidHelper.CompareGuidStringWithGuid(recipeId, fr.RecipeId));
         }
 
-        /// <summary>
-        ///  To get likes count for a Recipe
-        /// </summary>
+        /// <inheritdoc/>
         public async Task<int?> AllCountByRecipeIdAsync(string recipeId)
         {
             int? likes = await this.dbContext
@@ -40,10 +37,7 @@
             return likes;
         }
 
-        /// <summary>
-        ///  When a User Likes a Recipe, new record is created
-        /// </summary>        
-        /// 
+        /// <inheritdoc/>
         public async Task AddAsync(string userId, string recipeId)
         {
             FavouriteRecipe favouriteRecipe = new FavouriteRecipe()
@@ -57,9 +51,7 @@
             await this.dbContext.SaveChangesAsync();
         }
 
-        /// <summary>
-        /// When a User Unlikes a Recipe, the record is deleted
-        /// </summary>       
+        /// <inheritdoc/>
         public async Task DeleteAsync(string userId, string recipeId)
         {
             FavouriteRecipe? favouriteRecipe = await this.dbContext
@@ -75,9 +67,7 @@
             await this.dbContext.SaveChangesAsync();
         }
 
-        /// <summary>
-        ///  To get all Recipes, liked by a user
-        /// </summary>        
+        /// <inheritdoc/>    
         public async Task<ICollection<FavouriteRecipe>> GetAllByUserIdAsync(string userId)
         {
             ICollection<FavouriteRecipe> myRecipes = await this.dbContext
@@ -91,10 +81,7 @@
             return myRecipes;
         }
 
-        /// <summary>
-        /// To delete all records of User Liked Recipes (for user deleted profile, etc.)
-        /// </summary>
-        
+        /// <inheritdoc/>    
         public async Task DeleteAllByUserIdAsync(string userId)
         {
             var recipes = await this.dbContext 
@@ -102,14 +89,14 @@
                 .Where(fr => GuidHelper.CompareGuidStringWithGuid(userId, fr.UserId))
                 .ToListAsync();
 
-            this.dbContext.FavoriteRecipes.RemoveRange(recipes);
-            await this.dbContext.SaveChangesAsync();
+            if (recipes.Any())
+            {
+                this.dbContext.FavoriteRecipes.RemoveRange(recipes);
+                await this.dbContext.SaveChangesAsync();
+            }
         }
 
-        /// <summary>
-        /// To delete all likes when a Recipe is deleted
-        /// </summary>
-
+        /// <inheritdoc/>    
         public async Task DeleteAllByRecipeIdAsync(string recipeId)
         {
             var recipes = await this.dbContext
@@ -117,10 +104,12 @@
                 .Where(fr => GuidHelper.CompareGuidStringWithGuid(recipeId, fr.RecipeId))    
                 .ToListAsync();
 
-            this.dbContext.FavoriteRecipes.RemoveRange(recipes);
-            await this.dbContext.SaveChangesAsync();
+            if (recipes.Any())
+            {
+                this.dbContext.FavoriteRecipes.RemoveRange(recipes);
+                await this.dbContext.SaveChangesAsync();
+            }
         }
-
         
     }
 }
